improved input validation, added a validation lib and began memory optimization

+++ libs/libacpi/include/libacpi.h
@@ -25,6 +25,7 @@
 #include <stdbool.h>
 #include <libdevmgr.h>
 
+
 #if defined(__cplusplus)
 extern "C" {
 #endif
@@ -51,14 +52,14 @@ extern "C" {
 	//structure representing a ACPI header
 	struct acpi_header {
 		int8_t Signature[4];
-		unsigned long Length;
+		uint32_t Length;
 		int8_t Revision;
 		int8_t Checksum;
 		int8_t Oem_id[6];
 		int8_t Oem_table_id[8];
-		unsigned long Oem_revision;
+		uint32_t Oem_revision;
 		int8_t Creator_id[4];
-		unsigned long Creator_revision;
+		uint32_t Creator_revision;
 	};
 
 	//structure representing a ACPI subtable header
@@ -72,7 +73,7 @@ extern "C" {
 		struct acpi_subheader subheader;
 		int8_t acpi_processor_uid;
 		int8_t apic_id;
-		unsigned long flags; 
+		uint32_t flags; 
 	};
 
 	//structure representing a IOAPIC table
@@ -87,19 +88,19 @@ extern "C" {
 	//FADT info
 	int8_t PREFERED_PM_PROFILE;
 	int16_t SCI_INT;
-	unsigned long *SMI_CMD;
+	uint32_t *SMI_CMD;
 	int8_t ACPI_ENABLE;
 	int8_t ACPI_DISABLE;
 	int8_t S4BIOS_REQ;
 	int8_t PSTATE_CNT;
-	unsigned long *PM1a_EVT;
-	unsigned long *PM1b_EVT;
-	unsigned long *PM1a_CNT;
-	unsigned long *PM1b_CNT;
-	unsigned long *PM2_CNT;
-	unsigned long *PM_TMR;
-	unsigned long *GPE0;
-	unsigned long *GPE1;
+	uint32_t *PM1a_EVT;
+	uint32_t *PM1b_EVT;
+	uint32_t *PM1a_CNT;
+	uint32_t *PM1b_CNT;
+	uint32_t *PM2_CNT;
+	uint32_t *PM_TMR;
+	uint32_t *GPE0;
+	uint32_t *GPE1;
 	int8_t PM1_EVT_LEN;
 	int8_t PM1_CNT_LEN;
 	int8_t PM2_CNT_LEN;
@@ -117,7 +118,7 @@ extern "C" {
 	int8_t MON_ALRM;
 	int8_t CENTURY;
 	int16_t IAPC_BOOT_ARCH;
-	unsigned long FADT_FLAGS;
+	uint32_t FADT_FLAGS;
 	struct GAS_usable RESET_REG;
 	int8_t reset_value;
 
@@ -151,29 +152,29 @@ extern "C" {
 		int8_t CheckSum;
 		int8_t OemID[6];
 		int8_t Revision;
-	  	unsigned long *RsdtAddress;
+	  	uint32_t *RsdtAddress;
 	};
 
 	struct FACP {
 		struct acpi_header header;
-		unsigned long *FIRMWARE_CTRL;
-		unsigned long *DSDT;
+		uint32_t *FIRMWARE_CTRL;
+		uint32_t *DSDT;
 		int8_t reserved;
 		int8_t Preferred_PM_Profile;
 		int16_t SCI_INT;
-		unsigned long *SMI_CMD;
+		uint32_t *SMI_CMD;
 		int8_t ACPI_ENABLE;
 		int8_t ACPI_DISABLE;
 		int8_t S4BIOS_REQ;
 		int8_t PSTATE_CNT;
-		unsigned long *PM1a_EVT_BLK;
-		unsigned long *PM1b_EVT_BLK;
-		unsigned long *PM1a_CNT_BLK;
-		unsigned long *PM1b_CNT_BLK;
-		unsigned long *PM2_CNT_BLK;
-		unsigned long *PM_TMR_BLK;
-		unsigned long *GPE0_BLK;
-		unsigned long *GPE1_BLK;
+		uint32_t *PM1a_EVT_BLK;
+		uint32_t *PM1b_EVT_BLK;
+		uint32_t *PM1a_CNT_BLK;
+		uint32_t *PM1b_CNT_BLK;
+		uint32_t *PM2_CNT_BLK;
+		uint32_t *PM_TMR_BLK;
+		uint32_t *GPE0_BLK;
+		uint32_t *GPE1_BLK;
 		int8_t PM1_EVT_LEN;
 		int8_t PM1_CNT_LEN;
 		int8_t PM2_CNT_LEN;
@@ -193,7 +194,7 @@ extern "C" {
 		int8_t CENTURY;
 		int16_t IAPC_BOOT_ARCH;
 		int8_t Reserved2;
-		unsigned long Flags;
+		uint32_t Flags;
 		struct GAS reset_reg;
 		int8_t RESET_VALUE;
 		int8_t unneeded[244-131];
@@ -201,15 +202,15 @@ extern "C" {
 
 	struct FACS {
 		int8_t Signature[4];
-		unsigned long Length;
+		uint32_t Length;
 		int8_t Hardware_signature[4];
-		unsigned long* Firmware_waking_vector;
+		uint32_t* Firmware_waking_vector;
 	};
 	
 	struct MADT {
 		struct acpi_header header;
-		unsigned long *local_apic_address;
-		unsigned long flags;	
+		uint32_t *local_apic_address;
+		uint32_t flags;	
 	};
